refactor: move business logic from controllers to services
